Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/FileAPI/url/multi-global-origin-serialization.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/FileAPI/support/incumbent.sub.html
CONSOLE MESSAGE: line 19: TypeError: frames[0].createBlobURL is not a function. (In 'frames[0].createBlobURL()', 'frames[0].createBlobURL' is undefined)


FAIL Blob URL serialization (specifically the origin) in multi-global situations TypeError: frames[0].createBlobURL is not a function. (In 'frames[0].createBlobURL()', 'frames[0].createBlobURL' is undefined)
